Broke Panel and Tab out into there own files.  Bugfix for Button class.
git-svn-id: http://www.ardentheavyindustries.com/svn/syzygryd/trunk@54 1c6c7da2-4a7c-0410-9345-fc53533f7dce

+++ processing/controller_display/Button.pde
@@ -1,4 +1,7 @@
 /* -*- mode: java; c-basic-offset: 2; indent-tabs-mode: nil -*- */
+/**
+ * Button
+ */
 
 // NB: The following static methods ideally would be part of the
 // Button class, but processing turns classes into inner classes so
@@ -52,7 +55,7 @@ class Button
   PShape fullButton;
   PShape fullButtonActive, middleOnActive, middleActive, rightActive, leftActive;
 
-  Button(/*ArrayList _particleSystems,*/ int _x, int _y, float _sqLength, int _buttonSpacing, Panel _parent){
+  Button(/*ArrayList _particleSystems,*/ int _x, int _y, int _sqLength, int _buttonSpacing, Panel _parent){
  //   particleSystemsSimple = _particleSystems;
     x=_x;
     y=_y;
